requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
plotly==5.15.0
openpyxl==3.1.5
matplotlib==3.7.5
Pillow==10.2.0